backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
elevenlabs==1.5.0
anthropic==0.18.1
httpx==0.26.0

# ML dependencies for Wav2Lip
librosa==0.9.1
numpy==1.24.3
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
numba==0.57.1
